fix: success noti and config appointment page
