+++ src/api/appointment/index.ts
@@ -3,8 +3,12 @@ import CookieHelper, { CookieKeys } from 'src/utils/cookie-helper';
 import { Appointment, AppointmentDetail } from 'src/types/appointment';
 
 export class AppointmentApi {
-  static async getAppointments(request: {}): Promise<AppointmentDetail[]> {
-    const response = await apiGet('/appointments', getFormData(request));
-    return response;
+  // static async getAppointments(request: {}): Promise<AppointmentDetail[]> {
+  //   const response = await apiGet('/appointments', getFormData(request));
+  //   return response;
+  // }
+
+  static async getAppointments(params: Record<string, any>): Promise<any> {
+    return await apiGet('/api/v1/appointments', params);
   }
 }
