+++ src/contexts/appointment/appointment-context.tsx
@@ -8,28 +8,32 @@ import { AppointmentDetail } from 'src/types/appointment';
 import { useAuth } from '@hooks';
 
 interface ContextValue {
-  getListAppointmentsApi: UseFunctionReturnType<FormData, AppointmentDetail[]>;
+  // getListAppointmentsApi: UseFunctionReturnType<FormData, AppointmentDetail[]>;
+  getAppointmentListApi: UseFunctionReturnType<Record<string, any>, any>;
 }
 
 export const AppointmentContext = createContext<ContextValue>({
-  getListAppointmentsApi: DEFAULT_FUNCTION_RETURN
+  // getListAppointmentsApi: DEFAULT_FUNCTION_RETURN,
+  getAppointmentListApi: DEFAULT_FUNCTION_RETURN
 });
 
 const AppointmentProvider = ({ children }: { children: ReactNode }) => {
-  const getListAppointmentsApi = useFunction(AppointmentApi.getAppointments);
+  // const getListAppointmentsApi = useFunction(AppointmentApi.getAppointments);
+  const getAppointmentListApi = useFunction(AppointmentApi.getAppointments);
 
-  const { user } = useAuth();
+  // const { user } = useAuth();
 
-  useEffect(() => {
-    if (user?.role === 'ADMIN') {
-      getListAppointmentsApi.call(new FormData());
-    }
-  }, [user]);
+  // useEffect(() => {
+  //   if (user?.role === 'ADMIN') {
+  //     getListAppointmentsApi.call(new FormData());
+  //   }
+  // }, [user]);
 
   return (
     <AppointmentContext.Provider
       value={{
-        getListAppointmentsApi
+        // getListAppointmentsApi,
+        getAppointmentListApi
       }}
     >
       {children}
