+++ src/pages/_app.tsx
@@ -20,7 +20,6 @@ import { Header } from 'src/sections/header';
 import { Stack } from '@mui/material';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import { LocalizationProvider } from '@mui/x-date-pickers';
-import { Toaster } from 'react-hot-toast';
 
 const clientSideEmotionCache = createEmotionCache();
 
@@ -41,8 +40,6 @@ const App = (props: AppProps) => {
       <LocalizationProvider dateAdapter={AdapterDayjs}>
         <SnackbarProvider>
           <AuthProvider>
-            <Toaster />
-
             <AuthConsumer>
               {(auth) => {
                 const theme = createTheme(initialSettings);
