+++ src/sections/admin/appointment/index.tsx
@@ -1,0 +1,46 @@
+import { useState } from 'react';
+import { Box, InputAdornment, Stack, TextField } from '@mui/material';
+import ContentHeader from 'src/components/content-header';
+import AdvancedFilter from 'src/components/advanced-filter/advanced-filter';
+import { SearchIcon } from 'lucide-react';
+import { appointments } from 'src/utils/generate-mock';
+import AppointmentManagementList from 'src/sections/admin/appointment/appointment-management-list';
+
+export const AppointmentManagement: React.FC = () => {
+  const [searchInput, setSearchInput] = useState<string>('');
+  const handleSearch = () => {
+    console.log(searchInput);
+  };
+  return (
+    <>
+      <ContentHeader
+        title='Manage Appointments'
+        description='Showing: All Consultations of All Healthcare Providers'
+        rightSection={
+          <Stack direction={'row'} alignItems={'center'} gap={3} className='mt-4' flexWrap={'wrap'}>
+            <TextField
+              variant='outlined'
+              placeholder='t1faker@gmail.com'
+              className='max-sm:w-full w-[300px]'
+              value={searchInput}
+              onChange={(e) => setSearchInput(e.target.value)}
+              InputProps={{
+                endAdornment: (
+                  <InputAdornment position='end' className='cursor-pointer' onClick={handleSearch}>
+                    <SearchIcon />
+                  </InputAdornment>
+                )
+              }}
+            />
+            <Stack className='max-sm:ml-auto' direction={'row'} gap={1} alignItems={'center'}>
+              <AdvancedFilter filters={[]} />
+            </Stack>
+          </Stack>
+        }
+      />
+      <Box className='px-6 py-4'>
+        <AppointmentManagementList appointments={appointments} searchInput={searchInput} />
+      </Box>
+    </>
+  );
+};
